+++ app/sprinkles/admin/tests/Integration/Controller/UserControllerGuestTest.php
@@ -748,12 +748,12 @@ class UserControllerGuestTest extends TestCase
     /**
      * @return UserController
      */
-    private function getController()
+    protected function getController()
     {
         return new UserController($this->ci);
     }
 
-    private function setupUser()
+    protected function setupUser()
     {
         // Admin user, WILL have access
         $testUser = $this->createTestUser(false, true);