+++ app/sprinkles/admin/tests/Integration/Controller/UserControllerTest.php
@@ -1115,12 +1115,12 @@ class UserControllerTest extends TestCase
     /**
      * @return UserController
      */
-    private function getController()
+    protected function getController()
     {
         return new UserController($this->ci);
     }
 
-    private function setupUser()
+    protected function setupUser()
     {
         // Admin user, WILL have access
         $testUser = $this->createTestUser(true, true);